% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding_data.R
\name{encoding_data}
\alias{encoding_data}
\title{encoding_data}
\usage{
encoding_data(data)
}
\arguments{
\item{data}{A data frame obtained from get_data_from_bursa}
}
\value{No return value, used for encoding the data frame}
\description{
this function returns a data frame where character columns are encoded as UTF-8. it is a helper function in get_data_from_bursa
}
