% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_konya.R
\name{get_data_from_konya}
\alias{get_data_from_konya}
\title{get_data_from_bursa}
\usage{
get_data_from_konya(data_name)
}
\arguments{
\item{data_name}{A data set name obtained from list_data_names()}
}
\value{
A data frame that shows the observations for given data name
}
\description{
Given a data name, this function returns the data frame
}
\examples{
#NOT RUN
\donttest{
# It is just an example.
get_data_from_konya("bilgehane-merkez-konum-bilgileri")
}

}
