% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_names.R
\name{list_data_names}
\alias{list_data_names}
\title{list_data_names}
\usage{
list_data_names()
}
\value{
A character vector that shows the name of the data set
}
\description{
list the name of the data sets available in Bursa Open Data Portal
}
\examples{
\donttest{
list_data_names()
}
}
