% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_data_by_keyword.R
\name{search_data_by_keyword}
\alias{search_data_by_keyword}
\title{search_data_by_keyword}
\usage{
search_data_by_keyword(key_word)
}
\arguments{
\item{key_word}{A data set name which is a character object}
}
\value{
A character vector object that shows the name of the datasets including key word
}
\description{
Given a key word, this function returns the dataset names including key word
}
\examples{
\donttest{
search_data_by_keyword("ilce")
}
}
